requirements.txt
Django==5.2.1
djangorestframework==3.16.0
gunicorn==23.0.0
python-dotenv>=1.1.0
PyYAML==6.0.2
celery==5.5.2
redis==6.1.0
supervisor==4.2.5
mysqlclient==2.2.8
